yoga_assistant/pyproject.toml
[project]
name = "yoga-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=1.7.2",
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.6",
    "gunicorn>=23.0.0",
    "joblib>=1.5.3",
    "keras>=3.13.0",
    "langchain-google-genai>=4.1.2",
    "mediapipe>=0.10.31",
    "numpy>=2.4.0",
    "opencv-python-headless>=4.11.0.86",
    "passlib[bcrypt]>=1.7.4",
    "protobuf>=5.29.5",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "scikit-learn",
    "sqlalchemy>=2.0.45",
    "tensorflow-cpu>=2.20.0",
    "uvicorn>=0.40.0",
]
